Finish Call To Action Modal
Finished CTA Modal formatting and button styling.  I also got it to
line up with the header and first section borders by adding the same
amount of margin on the sides of all the content.  This is nice because
the button also has a flex quality so it will stay at the same sized
border.