Finish footer
I finished the footer, added some width to the thor quote so that it
doesn't get too large.  I want to investigate why there are seemingly
margins around the edges of the page.  Other than that I am as close
as I think I can reasonably be without getting nitpicky.