Finishing Touches
Added a 100vh height to the body and then made all the main content
sections flex:1 so that they would, theoretically, push the footer to
the bottom of the page.  However, it was already kind of there so I think
that this was just a formality.

Also noticed in dev tools that the <body> element had an 8px margin that
surrounded the entire page with a small white border.  I made a -8px margin
to offset this and it seems to have gone away.