Change hero text and embed youtube link.
Changed some of the text to actually represent that this page is about Goose.
Added some black gradiants to darken up the backgrounds, allowing the
white text to pop/be more visible.  Also added a youtube link and increased
some margins.  I've just noted out a lot of now irrelevant css instead of
deleting it so that I can keep it as a reference should I decide I don't like
certain things.

Next step is to add each band member's pictures and info to the card section!